+++ docs/.vitepress/config-en.ts
@@ -101,6 +101,7 @@ function sidebarFeatures(): DefaultTheme.SidebarItem[] {
         text: 'Dialog Panel',
         link: 'dialog-panel',
       },
+      { text: 'Fonts', link: 'fonts' },
       { text: 'Game Settings', link: 'game-settings' },
       { text: 'Gamepad support', link: 'gamepad' },
       {
@@ -113,6 +114,7 @@ function sidebarFeatures(): DefaultTheme.SidebarItem[] {
       },
       { text: 'Hotkeys', link: 'hotkeys' },
       { text: 'HUD Stats', link: 'hud-stats' },
+      { text: 'Input Prompts', link: 'input-prompts' },
       { text: 'Inventory', link: 'inventory' },
       { text: 'Items', link: 'items' },
       { text: 'Quests', link: 'quests' },
@@ -180,13 +182,18 @@ function commonCommandsSidebar(): DefaultTheme.SidebarItem[] {
 }
 
 function sidebarPlugins(): DefaultTheme.SidebarItem[] {
-  return [
-    {
-      text: 'Godot Narrat integration plugin',
-      link: '/plugins/godot/godot-plugin',
-    },
-    { text: 'Plugin API', link: '/plugins/plugins' },
-  ];
+  return {
+    text: 'Plugins',
+    collapsible: true,
+    items: [
+      {
+        text: 'Godot Narrat integration plugin',
+        link: '/plugins/godot/godot-plugin',
+      },
+      { text: 'Plugin API', link: '/plugins/plugins' },
+      { text: 'Plugin hooks', link: '/plugins/plugin-hooks' },
+    ],
+  };
 }
 
 function sidebarOthers(): DefaultTheme.SidebarItem[] {