+++ packages/narrat/src/components/game-splash/game-splash.vue
@@ -40,11 +40,11 @@ const gameLoaded = computed(() => main.loading.loaded);
 const gameTitle = computed(() => getCommonConfig().gameTitle || 'Narrat Game');
 
 function goToStartMenu() {
-  useScenes().changeScene('start-menu');
+  useScenes().goToStartMenuScene();
 }
 onMounted(() => {
   inputListener.value = useInputs().registerInputListener('game-splash', {
-    continue: {
+    confirm: {
       press: () => {
         goToStartMenu();
       },