+++ packages/narrat/src/config/common-config.ts
@@ -24,6 +24,8 @@ export const DialogPanelConfigSchema = Type.Optional(
     width: Type.Optional(Type.Number()),
     height: Type.Optional(Type.Number()),
     hideDuringTransition: Type.Optional(Type.Boolean()),
+    historyLength: Type.Optional(Type.Number()),
+    allowHistoryToggling: Type.Optional(Type.Boolean()),
     showAfterScriptEnd: Type.Optional(Type.Boolean()),
   }),
 );
@@ -127,13 +129,43 @@ export const SavesConfigSchema = Type.Object({
   slots: Type.Number(),
   runOnReload: Type.Optional(Type.String()),
   disabled: Type.Optional(Type.Boolean()),
+  autosaveDisabledOnLabels: Type.Optional(Type.Array(Type.String())),
+  autosaveFeedback: Type.Optional(
+    Type.Object({
+      enabled: Type.Boolean(),
+      text: Type.Optional(Type.String()),
+      backgroundImage: Type.Optional(Type.String()),
+      foregroundImage: Type.Optional(Type.String()),
+      duration: Type.Number(),
+    }),
+  ),
+  allowManualSave: Type.Optional(Type.Boolean()),
 });
 export type SavesConfig = Static<typeof SavesConfigSchema>;
+
+export const GraphicsSettingsSchema = Type.Object({
+  allowFullscreen: Type.Optional(Type.Boolean()),
+});
+export type GraphicsSettings = Static<typeof GraphicsSettingsSchema>;
 
 export const ScriptsConfigSchema = Type.Array(Type.String());
 export type ScriptsConfig = Static<typeof ScriptsConfigSchema>;
 
 export const defaultScriptsConfig = [];
+
+export const ScenesConfigSchema = Type.Object({
+  startMenuScene: Type.Optional(Type.String()),
+  gameScene: Type.Optional(Type.String()),
+  gameSplashScene: Type.Optional(Type.String()),
+});
+export type ScenesConfig = Static<typeof ScenesConfigSchema>;
+
+export const InputConfigSchema = Type.Object({
+  showPromptsOnKeyboard: Type.Optional(Type.Boolean()),
+  showPromptsOnGamepad: Type.Optional(Type.Boolean()),
+  showBottomLegend: Type.Optional(Type.Boolean()),
+});
+export type InputConfig = Static<typeof InputConfigSchema>;
 
 export const CommonConfigInputSchema = Type.Object({
   baseAssetsPath: Type.Optional(Type.String()),
@@ -143,6 +175,7 @@ export const CommonConfigInputSchema = Type.Object({
   images: Type.Optional(Type.Record(Type.String(), Type.String())),
   layout: LayoutConfigSchema,
   settings: Type.Optional(SettingsConfigSchema),
+  graphics: Type.Optional(GraphicsSettingsSchema),
   gameFlow: Type.Optional(
     Type.Object({
       labelToJumpOnScriptEnd: Type.Optional(Type.String()),
@@ -157,8 +190,11 @@ export const CommonConfigInputSchema = Type.Object({
   menuButtons: Type.Optional(MenuButtonsConfigSchema),
   debugging: Type.Optional(DebuggingConfigSchema),
   saves: Type.Optional(SavesConfigSchema),
+  input: Type.Optional(InputConfigSchema),
   hotkeys: HotkeysConfigSchema,
+  scenes: Type.Optional(ScenesConfigSchema),
 });
+
 export type CommonConfigInput = Static<typeof CommonConfigInputSchema>;
 
 export interface CommonConfig {
@@ -171,6 +207,7 @@ export interface CommonConfig {
   };
   layout: LayoutConfig;
   settings: SettingsConfig;
+  graphics: GraphicsSettings;
   gameFlow: {
     labelToJumpOnScriptEnd?: string;
   };
@@ -183,7 +220,9 @@ export interface CommonConfig {
   menuButtons: MenuButtonsConfig;
   debugging: DebuggingConfig;
   saves: SavesConfig;
+  input: InputConfig;
   hotkeys: HotkeysConfig;
+  scenes: ScenesConfig;
 }
 
 export const defaultCommonConfig: CommonConfig = {
@@ -194,6 +233,7 @@ export const defaultCommonConfig: CommonConfig = {
   images: {},
   layout: defaultLayoutConfig,
   settings: {},
+  graphics: {},
   gameFlow: {},
   dialogPanel: {
     overlayMode: true,
@@ -227,5 +267,11 @@ export const defaultCommonConfig: CommonConfig = {
     mode: 'manual',
     slots: 10,
   },
+  input: {
+    showPromptsOnKeyboard: true,
+    showPromptsOnGamepad: true,
+    showBottomLegend: true,
+  },
   hotkeys: {},
+  scenes: {},
 };