+++ packages/narrat/src/dialog-box.vue
@@ -58,6 +58,7 @@
             class="interact-button nrt-button override"
           >
             Continue
+            <InputPrompt input="continue" />
           </div>
         </div>
       </div>
@@ -84,6 +85,7 @@ import { InputListener, useInputs } from '@/stores/inputs-store';
 import { Interval, Timeout } from '@/utils/time-helpers';
 import { playLetterAudio, playDialogLineAudio } from '@/audio/audio-helpers';
 import { useVM } from '@/stores/vm-store';
+import InputPrompt from './components/input-prompt/input-prompt.vue';
 
 export interface TextAnimation {
   text: string;
@@ -111,6 +113,7 @@ const props = defineProps<{
   options: DialogBoxParameters;
   active: boolean;
   inputListener: InputListener | null;
+  index: number;
 }>();
 
 const navigation = ref<any | null>(null);
@@ -491,6 +494,13 @@ const dialogBoxClass = computed(() => {
   if (props.options.title) {
     css['dialog-box-followup'] = true;
   }
+  if (useDialogStore().isDialogCleared(props.index)) {
+    if (useDialogStore().clearedDialogVisible) {
+      css['dialog-box-cleared-visible'] = true;
+    } else {
+      css['dialog-box-cleared-disabled'] = true;
+    }
+  }
   if (props.options.old) {
     css['dialog-box-old'] = true;
   } else {
@@ -694,5 +704,11 @@ Otherwise hovering choices doesn't change the color of skill check prompts. */
 
 .dialog-box-old {
   opacity: 0.7;
+}
+.dialog-box-cleared-visible {
+  opacity: 0.6;
+}
+.dialog-box-cleared-disabled {
+  display: none;
 }
 </style>