+++ packages/narrat/src/examples/default/scripts/scripts.ts
@@ -11,7 +11,8 @@ import macros from './test_macros.narrat';
 import scenes from './test_scenes.narrat';
 import js from './test_js.narrat';
 import conditions from './conditions.narrat';
-
+import saves from './test_saves.narrat';
+import dialog from './test_dialog.narrat';
 export default [
   game,
   arrays,
@@ -26,4 +27,6 @@ export default [
   scenes,
   js,
   conditions,
+  saves,
+  dialog,
 ];