+++ packages/narrat/src/stores/achievements-store.ts
@@ -10,6 +10,7 @@ import {
   AchievementsConfig,
 } from '@/config/achievements-config';
 import { useConfig } from './config-store';
+import { updateGlobalSave } from '@/application/saving';
 
 export interface AchievementState {
   id: string;
@@ -74,14 +75,15 @@ export const useAchievements = defineStore('achievements', {
         const unlockTime = new Date().toISOString();
         existingAchievement.unlocked = true;
         existingAchievement.unlockTime = unlockTime;
-      }
-      if (getAchievementsConfig().notifyNewAchievements) {
-        const conf = getAchievementConfig(achievement);
-        useNotifications().addNotification(
-          `New Achievement: ${conf.name}`,
-          conf.description,
-          conf.icon,
-        );
+        if (getAchievementsConfig().notifyNewAchievements) {
+          const conf = getAchievementConfig(achievement);
+          useNotifications().addNotification(
+            `New Achievement: ${conf.name}`,
+            conf.description,
+            conf.icon,
+          );
+        }
+        updateGlobalSave();
       }
     },
   },